examples/sdn/manual-demo/ovs-plugin/dovesnap/dovesnap-main/pyproject.toml
[tool.poetry]
name = "dovesnap"
version = "1.1.22.dev"
description = "graphviz generator of dovesnap networks"
authors = ["Charlie Lewis <clewis@iqt.org>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
docker = "7.0.0"
faucetconfrpc = "0.55.67"
graphviz = "0.20.1"
"ruamel.yaml" = "^0.18.2"

[tool.poetry.dev-dependencies]
pytype = "2024.2.13"

[tool.poetry.scripts]
graph_dovesnap = "dovesnap.graph_dovesnap:main"

[tool.poetry.urls]
homepage = "https://github.com/IQTLabs/dovesnap"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
